Installation fix (#5)
Fixes Foreign key constraint

+++ database/migrations/2025_12_22_000001_create_default_folders_tables.php
@@ -21,7 +21,7 @@ return new class extends Migration
         if (!Schema::hasTable('enhanced_server_default_folder_servers')) {
             Schema::create('enhanced_server_default_folder_servers', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('default_folder_id');
+                $table->unsignedBigInteger('default_folder_id');
                 $table->unsignedInteger('server_id');
                 $table->integer('position')->default(0);
                 $table->timestamps();
@@ -38,4 +38,3 @@ return new class extends Migration
         Schema::dropIfExists('enhanced_server_default_folders');
     }
 };
-
